+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 - Use "hcw_" and "hcr_" prefixes in project API keys (#1148)
 - New integration: Google Chat (#423)
 - Fix the ping view to handle IPv4-mapped IPv6 addresses (#1169)
+- Update Dockerfile to listen on IPv6 as well (#1177)
 
 ## v3.10 - 2025-05-09